requirements.txt
asgiref==3.6.0
Django==4.1.7
django-modeltranslation==0.18.9
Pillow==9.4.0
sqlparse==0.4.3
typing_extensions==4.5.0
